Split reply-related sections of PostScraper into ReplyScraper
