+++ spec/services/post_scraper_spec.rb
@@ -30,8 +30,8 @@ RSpec.describe PostScraper do
     board = create(:board, creator: user)
 
     scraper = PostScraper.new(url, board.id)
-    allow(scraper).to receive(:prompt_for_user) { user }
-    allow(scraper).to receive(:set_from_icon).and_return(nil)
+    allow_any_instance_of(ReplyScraper).to receive(:prompt_for_user).and_return(user)
+    allow_any_instance_of(ReplyScraper).to receive(:set_from_icon).and_return(nil)
     expect(scraper.send(:logger)).to receive(:info).with("Importing thread 'linear b'")
 
     scraper.scrape!
@@ -54,8 +54,8 @@ RSpec.describe PostScraper do
     board = create(:board, creator: user)
 
     scraper = PostScraper.new(url, board.id)
-    allow(scraper).to receive(:prompt_for_user) { user }
-    allow(scraper).to receive(:set_from_icon).and_return(nil)
+    allow_any_instance_of(ReplyScraper).to receive(:prompt_for_user).and_return(user)
+    allow_any_instance_of(ReplyScraper).to receive(:set_from_icon).and_return(nil)
     expect(scraper.send(:logger)).to receive(:info).with("Importing thread 'linear b'")
 
     scraper.scrape!
@@ -100,14 +100,6 @@ RSpec.describe PostScraper do
     ])
   end
 
-  it "should raise an error when an unexpected character is found" do
-    url = 'http://wild-pegasus-appeared.dreamwidth.org/403.html?style=site&view=flat'
-    stub_fixture(url, 'scrape_no_replies')
-    scraper = PostScraper.new(url)
-    expect(scraper.send(:logger)).to receive(:info).with("Importing thread 'linear b'")
-    expect { scraper.scrape! }.to raise_error(UnrecognizedUsernameError)
-  end
-
   it "should raise an error when post is already imported" do
     board = create(:board)
     create(:character, screenname: 'wild_pegasus_appeared', user: board.creator)
@@ -130,27 +122,6 @@ RSpec.describe PostScraper do
     allow(scraper.send(:logger)).to receive(:info).with("Importing thread '#{new_title}'")
     expect { scraper.scrape! }.to raise_error(AlreadyImportedError)
     expect(Post.count).to eq(1)
-  end
-
-  it "should scrape character, user and icon properly" do
-    url = 'http://wild-pegasus-appeared.dreamwidth.org/403.html?style=site&view=flat'
-    stub_fixture(url, 'scrape_no_replies')
-    user = create(:user, username: "Marri")
-    board = create(:board, creator: user)
-
-    scraper = PostScraper.new(url, board.id, nil, nil, false, true)
-    allow(STDIN).to receive(:gets).and_return(user.username)
-    expect(scraper.send(:logger)).to receive(:info).with("Importing thread 'linear b'")
-    expect(scraper).to receive(:print).with("User ID or username for wild_pegasus_appeared? ")
-
-    scraper.scrape!
-
-    expect(Post.count).to eq(1)
-    expect(Reply.count).to eq(0)
-    expect(User.count).to eq(1)
-    expect(Icon.count).to eq(1)
-    expect(Character.count).to eq(1)
-    expect(Character.where(screenname: 'wild_pegasus_appeared').first).not_to be_nil
   end
 
   it "should only scrape specified threads if given" do
@@ -189,103 +160,6 @@ RSpec.describe PostScraper do
     expect(User.count).to eq(2)
     expect(Icon.count).to eq(30)
     expect(Character.count).to eq(8)
-  end
-
-  it "doesn't recreate characters and icons if they exist" do
-    url = 'http://wild-pegasus-appeared.dreamwidth.org/403.html?style=site&view=flat'
-    stub_fixture(url, 'scrape_no_replies')
-
-    user = create(:user, username: "Marri")
-    board = create(:board, creator: user)
-    nita = create(:character, user: user, screenname: 'wild_pegasus_appeared', name: 'Juanita')
-    icon = create(:icon, keyword: 'sad', url: 'http://v.dreamwidth.org/8517100/2343677', user: user)
-    gallery = create(:gallery, user: user)
-    gallery.icons << icon
-    nita.galleries << gallery
-
-    expect(User.count).to eq(1)
-    expect(Icon.count).to eq(1)
-    expect(Character.count).to eq(1)
-
-    scraper = PostScraper.new(url, board.id)
-    expect(scraper).not_to receive(:print).with("User ID or username for wild_pegasus_appeared? ")
-    expect(scraper.send(:logger)).to receive(:info).with("Importing thread 'linear b'") # just to quiet it
-
-    scraper.scrape!
-    expect(User.count).to eq(1)
-    expect(Icon.count).to eq(1)
-    expect(Character.count).to eq(1)
-  end
-
-  it "doesn't recreate icons if they already exist for that character with new urls" do
-    url = 'http://wild-pegasus-appeared.dreamwidth.org/403.html?style=site&view=flat'
-    stub_fixture(url, 'scrape_no_replies')
-
-    user = create(:user, username: "Marri")
-    board = create(:board, creator: user)
-    nita = create(:character, user: user, screenname: 'wild_pegasus_appeared', name: 'Juanita')
-    icon = create(:icon, keyword: 'sad', url: 'http://glowfic.com/uploaded/icon.png', user: user)
-    gallery = create(:gallery, user: user)
-    gallery.icons << icon
-    nita.galleries << gallery
-
-    expect(User.count).to eq(1)
-    expect(Icon.count).to eq(1)
-    expect(Character.count).to eq(1)
-
-    scraper = PostScraper.new(url, board.id)
-    expect(scraper).not_to receive(:print).with("User ID or username for wild_pegasus_appeared? ")
-    expect(scraper.send(:logger)).to receive(:info).with("Importing thread 'linear b'") # just to quiet it
-
-    scraper.scrape!
-    expect(User.count).to eq(1)
-    expect(Icon.count).to eq(1)
-    expect(Character.count).to eq(1)
-  end
-
-  it "handles Kappa icons" do
-    kappa = create(:user, id: 3)
-    char = create(:character, user: kappa)
-    gallery = create(:gallery, user: kappa)
-    char.galleries << gallery
-    icon = create(:icon, user: kappa, keyword: '⑮ mountains')
-    gallery.icons << icon
-    tag = build(:reply, user: kappa, character: char)
-    expect(tag.icon_id).to be_nil
-    scraper = PostScraper.new('')
-    scraper.send(:set_from_icon, tag, 'http://irrelevanturl.com', 'f.1 mountains')
-    expect(Icon.count).to eq(1)
-    expect(tag.icon_id).to eq(icon.id)
-  end
-
-  it "handles icons with descriptions" do
-    user = create(:user)
-    char = create(:character, user: user)
-    gallery = create(:gallery, user: user)
-    char.galleries << gallery
-    icon = create(:icon, user: user, keyword: 'keyword blah')
-    gallery.icons << icon
-    tag = build(:reply, user: user, character: char)
-    expect(tag.icon_id).to be_nil
-    scraper = PostScraper.new('')
-    scraper.send(:set_from_icon, tag, 'http://irrelevanturl.com', 'keyword blah (Accessbility description.)')
-    expect(Icon.count).to eq(1)
-    expect(tag.icon_id).to eq(icon.id)
-  end
-
-  it "handles kappa icons with descriptions" do
-    kappa = create(:user, id: 3)
-    char = create(:character, user: kappa)
-    gallery = create(:gallery, user: kappa)
-    char.galleries << gallery
-    icon = create(:icon, user: kappa, keyword: '⑮ keyword blah')
-    gallery.icons << icon
-    tag = build(:reply, user: kappa, character: char)
-    expect(tag.icon_id).to be_nil
-    scraper = PostScraper.new('')
-    scraper.send(:set_from_icon, tag, 'http://irrelevanturl.com', 'f.1 keyword blah (Accessbility description.)')
-    expect(Icon.count).to eq(1)
-    expect(tag.icon_id).to eq(icon.id)
   end
 
   it "can fail a download" do
